Moved SwitchPreference to preference package
